# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mini eCommerce API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: health
      responses:
        '200': { description: OK }
  /login:
    post:
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        '401': { description: Unauthorized }
  /products:
    get:
      summary: List products
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: inStock
          schema: { type: string, enum: ['1','0','true','false'] }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 10 }
      responses:
        '200': { description: OK }
    post:
      summary: Create product (admin)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, category, inStock]
              properties:
                name: { type: string }
                price: { type: number }
                category: { type: string }
                inStock: { type: boolean }
      responses:
        '201': { description: Created }
        '401': { description: Unauthorized }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT